# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github.com_sonr-io_speedway_server_routes.CAResponseBody:
    properties:
      address:
        type: string
    type: object
...
  github.com_sonr-io_speedway_server_routes.FailedResponse:
    properties:
      error:
        type: string
    type: object
...
  server_routes.CreateSchemaResponse:
    properties:
      definition:
        $ref: '#/definitions/types.SchemaDefinition'
      whatIs:
        $ref: '#/definitions/types.WhatIs'
    type: object
  server_routes.FailedLogin:
    properties:
      success:
        type: boolean
    type: object
  server_routes.FailedResponse:
    properties:
      error:
        type: string
    type: object
...
  server_routes.SuccessfulLogin:
    properties:
      address:
        type: string
      success:
        type: boolean
    type: object
...
  types.Alias:
    properties:
      amount:
        description: Amount is the amount listed for purchasing the Alias from the
          User/Application
        type: integer
      is_for_sale:
        description: IsForSale is the boolean value indicating if the Alias is for
          sale
        type: boolean
      name:
        description: Name is the string name of an Alias
        type: string
    type: object
...
  types.DIDDocument:
    properties:
      also_known_as:
        items:
          type: string
        type: array
      assertion_method:
        items:
          type: string
        type: array
      authentication:
        items:
          type: string
        type: array
      capability_delegation:
        items:
          type: string
        type: array
      capability_invocation:
        items:
          type: string
        type: array
      context:
        items:
          type: string
        type: array
      controller:
        items:
          type: string
        type: array
      id:
        type: string
      key_agreement:
        items:
          type: string
        type: array
      service:
        items:
          $ref: '#/definitions/types.Service'
        type: array
      verification_method:
        items:
          $ref: '#/definitions/types.VerificationMethod'
        type: array
    type: object
  types.KeyValuePair:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  types.MsgTransferAliasResponse:
    properties:
      success:
        description: Success is true if the Alias was successfully transferred.
        type: boolean
      who_is:
        $ref: '#/definitions/types.WhoIs'
        description: WhoIs is the updated WhoIs, contains the DID document and associated
          metadata.
    type: object
  types.SchemaDefinition:
    properties:
      creator:
        type: string
      did:
        description: |-
          Represents the types of fields a schema can have
          the DID for this schema
        type: string
      fields:
        description: the properties of this schema
        items:
          $ref: '#/definitions/types.SchemaKindDefinition'
        type: array
      label:
        description: an alternative reference point
        type: string
    type: object
  types.SchemaKindDefinition:
    properties:
      field:
        description: Type of a single schema property
        type: integer
      link:
        type: string
      link_kind:
        description: Optional field for a link context if `SchemaKind` is of type
          `Link`
        type: integer
      name:
        type: string
    type: object
  types.SchemaReference:
    properties:
      cid:
        description: a reference to information stored within an IPFS node.
        type: string
      did:
        description: the DID for this schema
        type: string
      label:
        description: an alternative reference point
        type: string
    type: object
  types.Service:
    properties:
      id:
        type: string
      service_endpoint:
        items:
          $ref: '#/definitions/types.KeyValuePair'
        type: array
      type:
        type: string
    type: object
  types.VerificationMethod:
    properties:
      controller:
        type: string
      credential_json:
        items:
          type: integer
        type: array
      id:
        type: string
      public_key_base58:
        type: string
      public_key_jwk:
        items:
          $ref: '#/definitions/types.KeyValuePair'
        type: array
      type:
        type: string
    type: object
  types.WhatIs:
    properties:
      creator:
        description: Creator is the DID of the creator
        type: string
      did:
        description: DID is the DID of the object
        type: string
      is_active:
        description: IsActive is the status of the DID Document
        type: boolean
      metadata:
        additionalProperties:
          type: string
        description: Metadata is a map of key-value pairs that can be used to store
          additional information about the WhatIs (Schema)
        type: object
      schema:
        $ref: '#/definitions/types.SchemaReference'
        description: Object_doc is the object document
      timestamp:
        description: Timestamp is the time of the last update of the DID Document
        type: integer
    type: object
  types.WhoIs:
    properties:
      alias:
        description: Alias is the list of registered `alsoKnownAs` identifiers of
          the User or Application
        items:
          $ref: '#/definitions/types.Alias'
        type: array
      controllers:
        description: Credentials are the biometric info of the registered name and
          account encoded with public key
        items:
          type: string
        type: array
      did_document:
        $ref: '#/definitions/types.DIDDocument'
        description: DIDDocument is the bytes representation of DIDDocument within
          the WhoIs. Initially marshalled as JSON.
      is_active:
        description: IsActive is the status of the DID Document
        type: boolean
      metadata:
        additionalProperties:
          type: string
        description: Metadata is a map of key-value pairs that can be used to store
          additional information about the DID Document
        type: object
      owner:
        description: Owner is the top level DID of the User or Application derived
          from the multisignature wallet.
        type: string
      timestamp:
        description: Timestamp is the time of the last update of the DID Document
        type: integer
      type:
        description: 'Type is the kind of the entity. Possible values are: "user",
          "application"'
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: team@sonr.io
    name: API Support
    url: https://sonr.io/
  description: Create accounts, schemas, buckets and objects in a scalable way utilizing
    the Sonr architecture.
  license:
    name: OpenGLv3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  termsOfService: https://sonr.io/terms
  title: Speedway API
  version: "2.0"
paths:
  /account/create:
    post:
      consumes:
      - application/json
      description: Create a new account on Sonr using the Registry module of Sonr's
        Blockchain.
      parameters:
      - description: password
        example: '"Password"'
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_sonr-io_speedway_server_routes.CAResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_sonr-io_speedway_server_routes.FailedResponse'
      summary: CreateAccount
      tags:
      - Account
...
  /account/login:
    post:
      consumes:
      - application/json
      description: Login to an existing account on Sonr using the Registry module
        of Sonr's Blockchain.
      parameters:
      - description: did
        example: '"snr172ljvam8m7xxlv59v6w27lula58zwwct3vgn9p"'
        in: body
        name: Did
        required: true
        schema:
          type: string
      - description: password
        example: '"Password"'
        in: body
        name: Password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_routes.SuccessfulLogin'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_routes.FailedLogin'
      summary: LoginAccount
      tags:
      - Account
...
  /alias/transfer:
    post:
      consumes:
      - application/json
      description: Transfer an alias
      parameters:
      - description: Creator
        in: body
        name: creator
        required: true
        schema:
          type: string
      - description: Alias
        in: body
        name: alias
        required: true
        schema:
          type: string
      - description: Recipient
        in: body
        name: recipient
        required: true
        schema:
          type: string
      - description: Amount
        in: body
        name: amount
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MsgTransferAliasResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_routes.FailedResponse'
      summary: TransferAlias
      tags:
      - Alias
...
  /schema/create:
    post:
      consumes:
      - application/json
      description: Create a schema utilizing motor client. Returns the WhatIs of the
        schema created.
      parameters:
      - description: Label of the schema
        in: query
        name: label
        required: true
        type: string
      - description: Fields of the schema
        in: query
        name: fields
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_routes.CreateSchemaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_routes.FailedResponse'
      summary: CreateSchema
      tags:
      - Schema
...
swagger: "2.0"